more modifications

+++ db/migrate/20180107125746_create_chequeos.rb
@@ -2,7 +2,7 @@ class CreateChequeos < ActiveRecord::Migration[5.1]
   def change
     create_table :chequeos do |t|
       t.string :fecha
-      t.integer :peso
+      t.decimal :peso
       t.references :user, foreign_key: true
 
       t.timestamps
